+++ packaging.R
@@ -25,12 +25,16 @@ create_package(
     URL = "https://github.com/mayer79/splitTools",
     BugReports = "https://github.com/mayer79/splitTools/issues",
     Depends = "R (>= 3.5.0)",
+    VignetteBuilder = "knitr",
     License = "GPL(>= 2)",
     Maintainer = "Michael Mayer <mayermichael79@gmail.com>"))
 
 file.copy(file.path(pkg, "DESCRIPTION"), to = getwd(), overwrite = TRUE)
 # Use package has no option to look for pkg, so we first copy description from pkg, modify it and move back
 use_package("stats", "Imports")
+use_package("knitr", "Suggests")
+use_package("ranger", "Suggests")
+
 # use_rcpp(pkg)
 
 # Set up other files -------------------------------------------------
@@ -45,17 +49,24 @@ file.copy(c("NEWS.md", "README.md", "cran-comments.md", "DESCRIPTION", ".Rbuildi
 file.copy(list.files("R", full.names = TRUE), file.path(pkg, "R"), overwrite = TRUE)
 devtools::document(pkg)
 
+# Copy vignette
+# use_vignette(name = "splitTools", title = "splitTools")
+dir.create(file.path(pkg, "vignettes"))
+dir.create(file.path(pkg, "doc"))
+dir.create(file.path(pkg, "Meta"))
+file.copy(list.files("vignettes", full.names = TRUE),
+          file.path(pkg, "vignettes"), overwrite = TRUE)
+
+devtools::build_vignettes(pkg)
+
 # Check
 check(pkg, manual = TRUE)
 
 # Create
 build(pkg)
-build(pkg, binary = TRUE)
 
 # Install
 install(pkg)
-
-# modify .Rbuildignore in build project to ignore the proj file.
 
 check_win_devel(pkg)